requirements.txt
flask>=3.0.0
flask-cors>=6.0.0
pymongo>=4.0.0
openai>=1.0.0
tiktoken>=0.5.0
markdown>=3.5.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
groq>=0.4.0
chromadb>=0.4.0
requests>=2.31.0
tqdm>=4.66.0

